+++ liquid/golden/illegal_tag.py
@@ -1,0 +1,11 @@
+"""Golden tests cases for testing illegal tags."""
+
+from liquid.golden.case import Case
+
+cases = [
+    Case(
+        description="unknown tag",
+        template=r"{% nosuchthing %}",
+        expect="",
+    ),
+]